Add files via upload
The two main files in this project are the Selenium.py file and the file where I perform cleaning and visualisation in Jupyter Notebook. I also produced a separate file for webscraping the news for 13 January. There are .csv files for separate days where I scraped for the global breaking news (Jan 12 and Jan 18, 2024.)